# [WIP]: Part 2: Improve performance of PerWikiCourseStats#stats