@gabina 2025-08-08
> Explicitly filter by namespace: 0 (main article namespace).
> 
> Improved accuracy: Articles outside of the main namespace (e.g., Talk, User pages) was not be included in the count of course-edited articles.

According to what I understand, this PR changes the behavior for the CSV report. Before this PR, we counted articles edited based on all articles updated during the course. After this PR, we only count articles edited on mainspace. Am I understanding correctly? There are a lot of courses that track not only the namespace but other ones. For example, course [32690](https://outreachdashboard.wmflabs.org/courses/Tyap_wikimedians_Community/AWC_2025_in_the_Tyap_Community). I'm not sure if counting only articles in nampesace is the expected behavior (or more accurate), but you should check with Sage. It could be helpful if you generate both report versions and compare results for a course that tracks several namespaces.