@Abishekcs 2025-08-08
> Am I understanding correctly?

Yes,

> I'm not sure if counting only articles in nampesace is the expected behavior (or more accurate), but you should check with Sage. It could be helpful if you generate both report versions and compare results for a course that tracks several namespaces.

Okay, i will try to check with more namespaces. Thank you for the review!