@ragesoss 2025-08-11
@Abishekcs Gabina is correct that the expected behavior would be to include all tracked namespaces, in cases where it's not just the default of mainspace for that course.