@Abishekcs 2025-08-14
Closing this PR as PR #6432 and #6433 optimize it by not requiring to use `distinct` anymore.